Rename local functions to common.
