+++ book/models.py
@@ -4,7 +4,7 @@ from django.contrib.sites.managers import CurrentSiteManager
 from django.utils.text import slugify
 from django.conf import settings
 from django.urls import reverse
-from local import functions
+from common import functions
 
 # Functions for image path upload
 def volume_directory_location(instance, filename):
